feat: Join Lobby screen code only
