+++ app/src/main/AndroidManifest.xml
@@ -17,6 +17,9 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
+            android:name=".activities.JoinLobbyActivity"
+            android:exported="false" />
+        <activity
             android:name=".activities.CreateLobbyActivity"
             android:exported="false" />
         <activity
