+++ app/src/main/res/layout/activity_main.xml
@@ -38,7 +38,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="16dp"
         android:onClick="btnCLobbyClicked"
-        android:text="@string/tmpCreateLobby"
+        android:text="@string/createLobby"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.497"
         app:layout_constraintStart_toStartOf="parent"
@@ -49,8 +49,8 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="80dp"
-        android:onClick="tmpBtnJLobbyClicked"
-        android:text="@string/tmpJoinLobby"
+        android:onClick="btnJLobbyClicked"
+        android:text="@string/joinLobby"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.498"
         app:layout_constraintStart_toStartOf="parent"
