+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,8 @@
 <resources>
     <string name="appName">Wizard</string>
     <!--MainActivity-->
-    <string name="tmpCreateLobby">Create Lobby</string>
-    <string name="tmpJoinLobby">TMP Join Lobby</string>
+    <string name="createLobby">Create Lobby</string>
+    <string name="joinLobby">Join Lobby</string>
     <string name="tmpGameActivity">Game Activity</string>
     <!--RegisterActivity-->
     <string name="register">Register</string>
@@ -44,8 +44,10 @@
     <string name="gameNoPoints">00</string>
     <!--CreateLobbyActivity-->
     <string name="createLobbyText">Create Lobby</string>
-    <string name="createLobbyMenu" >Back</string>
+    <string name="createLobbyMenu">Back</string>
     <string name="createLobbyCreate">Create</string>
     <string name="createLobbyIsPublic">Public?</string>
     <string name="createLobbyMaxPlayerText">Maximum number of players:</string>
+    <string name="joinLobbyText">Join Lobby</string>
+    <string name="joinLobbyLobbyCode">Join Lobby with code</string>
 </resources>
